:green_heart::checkered_flag: Another easy fix for Windows CI

+++ utils/products/lua/checkout.py
@@ -73,4 +73,5 @@ def get_dependency():
     shell.rm(destination)
     move_files()
 
-    shell.rmtree(os.path.join(ANTHEM_SOURCE_ROOT, product.key, "temp"))
+    if not platform.system() == "Windows" and not data.build.ci:
+        shell.rmtree(os.path.join(ANTHEM_SOURCE_ROOT, product.key, "temp"))
